+++ src/store/index.ts
@@ -8,3 +8,5 @@ export default pinia
 export * from './modules/app'
 export * from './modules/user'
 export * from './modules/theme'
+export * from './modules/tags-view'
+export * from './modules/dict'